+++ upload_controller_service/adapters/inbound/kafka_consume.py
