+++ upload_controller_service/adapters/outbound/kafka_produce.py
